Use token in pkcs11 URLs
It turns out that `pkcs11:id=x` wasn't being matched and libp11 would
select the last valid ID. This would be the highest numbered I2C bus
(i.e., the last one checked). Out of an amazing amount of luck,
NervesKeys were being put on the highest numbered I2C bus apparently.

The fix is to specify `pkcs11:token=x`.

See the code snippet here for the match logic:

https://github.com/OpenSC/libp11/blob/6064a0a8fef06d96f3e6e13f02b2f0513aa93214/src/eng_back.c#L473-L483

The word `token` maps to `match_tok->label`. The ID isn't checked in
the loop so all slots match. This library makes a slot per I2C bus and
they're made in order. The last match wins in that code. Therefore, if
you have `i2c-0` and `i2c-1`, the ATECC608A had to be on `i2c-1` to
work. Now it can be on `i2c-0`.