Fix segfault with libp11 v0.4.11
For reasons that aren't clear to me libp11 v0.4.11 has started probing
the ECC signature length. libp11 v0.4.10 does not do this. A probe
consists of passing a NULL signature into C_Sign(). Previously this code
returned success - most likely it was copy-paste and it seemed
reasonable. However, now that libp11 probes the length with a NULL
signature, the success response caused a code path in libp11 to be
followed that deferenced the NULL pointer. Returning a signature too
short error causes the right code path to be followed.