Update nerves_key_pkcs11.c
Changed the .ulMaxSessionCount to 4 (rather than 1)  so that when the "{:ok, engine} = NervesKey.PKCS11.load_engine()" is called more than once, it doesn't return "** (MatchError) no match of right hand side value: {:error, :ctrl_cmd_failed}". 

In several cases, the user may require to call the engine multiple times. 

One other option is to create it as a gen server, but this hotfix will fix the {:error, :ctrl_cmd_failed}.